ui-charts: make useCanvas completely independent od SpaceTimeChart
Move some utils functions, hooks and types from /spaceTimeChart into /common.
