+++ front/ui/ui-charts/src/common/helpers/png.ts
