+++ front/ui/ui-charts/src/common/index.ts
@@ -1,3 +1,9 @@
 export { usePicking, useDraw } from './useCanvas';
 
-export type { HoveredItem, DrawingFunction, PickingDrawingFunction, PickingElement } from './types';
+export type {
+  HoveredItem,
+  DrawingFunction,
+  PickingDrawingFunction,
+  PickingElement,
+  Point,
+} from './types';
