+++ front/ui/ui-charts/src/common/useDevicePixelRatio.ts
