+++ front/ui/ui-charts/src/common/useSize.ts
@@ -2,7 +2,6 @@ import { useEffect, useState } from 'react';
 
 /**
  * This hook takes a DOM element reference as input, and returns always updated width and height.
- * It is an internal hook, and should only be used inside SpaceTimeChart.
  */
 export function useSize(dom?: HTMLElement | null) {
   const [size, setSize] = useState<{ width: number; height: number }>({
