+++ front/ui/ui-charts/src/spaceTimeChart/components/ConflictLayer.tsx
@@ -1,11 +1,11 @@
 import { useCallback } from 'react';
 
+import { hexToRgb, indexToColor } from '../../common/helpers/colors';
 import type { DrawingFunction, PickingDrawingFunction, PickingElement } from '../../common/types';
 import { useDraw, usePicking } from '../../common/useCanvas';
 import { SpaceTimeChartCanvasContext } from '../lib/context';
 import type { SpaceTimeChartContextType } from '../lib/types';
 import { drawAliasedRect } from '../utils/canvas';
-import { indexToColor, hexToRgb } from '../utils/colors';
 
 export type Conflict = {
   timeStart: number;
