+++ front/ui/ui-charts/src/spaceTimeChart/components/SpaceTimeChart.tsx
@@ -7,9 +7,9 @@ import { TimeCaptions } from './TimeCaptions';
 import TimeGraduations from './TimeGraduations';
 import type { PickingElement } from '../../common/types';
 import { useCanvas } from '../../common/useCanvas';
+import { useSize } from '../../common/useSize';
 import { useMouseInteractions } from '../hooks/useMouseInteractions';
 import { useMouseTracking } from '../hooks/useMouseTracking';
-import { useSize } from '../hooks/useSize';
 import { DEFAULT_THEME } from '../lib/consts';
 import { MouseContext, SpaceTimeChartCanvasContext, SpaceTimeChartContext } from '../lib/context';
 import { validateTheme } from '../lib/theme';
