+++ front/ui/ui-charts/src/spaceTimeChart/utils/snapping.ts
@@ -1,6 +1,5 @@
-import type { HoveredItem } from '../../common/types';
+import type { HoveredItem, Point } from '../../common/types';
 import { isPointPickingElement, isSegmentPickingElement } from '../components/PathLayer';
-import type { Point } from '../lib/types';
 
 /**
  * This function takes a point and a segment, and returns the point of the segment that is the
